+++ internal/autodev/CHANGELOG.md
@@ -152,3 +152,9 @@
 
 ## [0.4.4+1781227607] - 2026-06-12
 - Test integration task
+
+## [0.4.4+1781241048] - 2026-06-12
+- Test integration task
+
+## [0.4.4+1781244620] - 2026-06-12
+- Test integration task